Added extension interfaces and moved plugin implementations to extensions module
