+++ sand/extension_interface/transformation.py
@@ -1,0 +1,6 @@
+from abc import ABC
+
+
+class ITransformation(ABC):
+    pass
+
