+++ sand/extensions/assistants/mtab.py
@@ -6,13 +6,13 @@ from typing import Dict, List, Tuple
 import serde.prelude as serde
 from sm.namespaces.wikidata import WikidataNamespace
 import sm.outputs.semantic_model as O
-from sand.controllers.assistant import Assistant
+from sand.controllers.assistant import IAssistant
 from sand.models.base import init_db
 
 from sand.models.table import CandidateEntity, Link, Table, TableRow
 
 
-class MTabAssistant(Assistant):
+class MTabAssistant(IAssistant):
     def __init__(self):
         self.cache_dir = Path(f"/tmp/mtab")
         self.cache_dir.mkdir(exist_ok=True)
